Clean, fix, result  - Made some cleanup and formatting  - Protecting controllers for future bugs  - Added json result for methods

+++ src/Controller/MachineController.php
@@ -19,27 +19,53 @@ class MachineController extends AbstractController
     {
         $m = new Machine();
         $j = json_decode($r->getContent(), true);
+        if (!ctype_digit($j['memory'] or !ctype_digit($j['cpus'])))
+        {
+            return $this->json(null, 400);
+        }
+
         $m->setMemory($j['memory'])->setCpus($j['cpus']);
         $lb->findProcessForMachine($m);
         $em->persist($m);
         $em->flush();
+
+        $pid = $m->getProcess() === null ? null : $m->getProcess()->getId();
         return $this->json([
+            'process_id' => $pid,
+            'machine_id' => $m->getId(),
         ]);
     }
 
     #[Route('/remove', name: 'remove_machine')]
     public function remove(Request $r, LoadBalancer $lb, EntityManagerInterface $em, MachineRepository $mrep): JsonResponse
     {
+        $mid = null;
+        $pid = null;
+
         $j = json_decode($r->getContent(), true);
+        if (!ctype_digit($j['id']))
+        {
+            return $this->json(null, 400);
+        }
+
         $m = $mrep->find($j['id']);
         $p = $m->getProcess();
         $em->beginTransaction();
         $em->remove($m);
         $em->flush();
-        $lb->findMachineForProcess($p);
+        if (!is_null($p))
+        {
+            $lb->findMachineForProcess($p);
+            $mid = $p->getMachine() === null ? null : $p->getMachine()->getId();
+            $pid = $p->getId();
+        }
+
         $em->flush();
         $em->commit();
+
         return $this->json([
+            'process_id' => $pid,
+            'machine_id' => $mid,
         ]);
     }
 }
